% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcObject.R
\name{convergence.test}
\alias{convergence.test}
\title{Convergence Test}
\usage{
convergence.test(object, nsamples = 10, frac1 = 0.1, frac2 = 0.5,
  thin = 1, plot = FALSE, ...)
}
\arguments{
\item{object}{}

\item{nsamples}{}

\item{frac1}{}

\item{frac2}{}

\item{thin}{}

\item{plot}{}

\item{...}{}
}
\value{

}
\description{
\code{convergence.test}
}
\details{
\code{convergence.test}
}

