% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{extractBaseInfo}
\alias{extractBaseInfo}
\title{Extract Basic Information From Parameter Object}
\usage{
extractBaseInfo(parameter)
}
\arguments{
\item{parameter}{A parameter object that corrosponds to
one of the model types, such as "ROC", or "FONSE".}
}
\value{
varList Returns a list of the variables extracted from the parameter
object.
}
\description{
\code{extractBaseInfo} will extract all of the shared information
between all known parameter object types and store them in R data types. It 
returns a list of these variables.
}
\details{
This function calls the C++ getter functions to get all of the common
parameter information and store them in R datastructures. The purpose of this
is to be able to use R's save function to serialize objects, which Rcpp does
not support with C++ classes.
}

