% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{getCodonCountsForAA}
\alias{getCodonCountsForAA}
\title{Get Codon Counts For Each Amino Acid}
\usage{
getCodonCountsForAA(aa, genome)
}
\arguments{
\item{aa}{A one character representation of an amino acid.}

\item{genome}{A genome object from which the counts of each
codon can be obtained.}
}
\value{
codonCounts Returns a matrix storing the codonCounts.
}
\description{
\code{getCodonCountsForAA} returns a matrix filled with 
the number of times a codon is seen in each gene.
}
\details{
The returned matrix will have the row corrospond to the
genes in the genome and the columns corrospond to the codons for the 
given aa. The values will the number of times the codon is present in 
that gene.
}

