% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{initializeFONSEParameterObject}
\alias{initializeFONSEParameterObject}
\title{Initialize  FONSE Parameter}
\usage{
initializeFONSEParameterObject(genome, sphi, numMixtures, geneAssignment,
  expressionValues = NULL, split.serine = TRUE,
  mixture.definition = "allUnique", mixture.definition.matrix = NULL)
}
\arguments{
\item{genome}{A genome object which the parameter use.}

\item{sphi}{Initial sphi values that corrosponds with the sphi for
each mixture. sphi is a vector whose length should equal numMixtures.}

\item{numMixtures}{The number of mixtures the parameter 
should use with the genome. This should be a positive number.}

\item{geneAssignment}{A vector holding that corrosponds to each
gene in the genome. The vector size should equal the number of 
genes in the genome. The assignment is to which mixture the 
gene initially starts in. Valid values for the vector go from
1 to numMixtures.}

\item{expressionValues}{A vector containing the starting phi 
values. This is an optional argument and is set to null if
no value is given. If it is provided, the length of the 
vector should equal the number of genes in the genome.}

\item{split.serine}{Whether serine should be considered as 
one or two amino acids when running the model. TRUE and FALSE
are the only valid values. The default value for split.serine is
TRUE.}

\item{mixture.definition}{A string describing how each mixture should
be treated with respect to mutation and selection. Valid values consist
of "allUnique", "mutationShared", and "selectionShared". The default value
for mixture.definition is "allUnique". See details for more information.}

\item{mixture.definition.matrix}{A matrix representation of how
the mutation and selection categories corrospond to the mixtures.
The default value for mixture.definition.matrix is NULL. If provided,
the model will use the matrix to initialize the mutation and selection
categories instead of the definition listed directly above. See details
for more information.}
}
\value{
parameter Returns an initialized FONSE parameter object.
}
\description{
\code{initializeFONSEParameterObject} will initialize
the ROC Parameter object with the given values.
}
\details{
\code{initializeFONSEParameterObject} should never be called; it
is called from the \code{initializeParameterObject}. Because of this,
no error checking is doen in this function. Additional documentation on 
the mixture.definition.matrix and mixture.definition can
be found in the documentation for \code{initializeParameterObject}.
}

