% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genomeObject.R
\name{initializeGenomeObject}
\alias{initializeGenomeObject}
\title{Genome Initialization}
\usage{
initializeGenomeObject(file, fasta = TRUE, expression.file = NULL,
  append = FALSE)
}
\arguments{
\item{file}{A string containing the location of the input file.}

\item{fasta}{A boolean value which decides whether to initialize with a
fasta file or an RFP value file. (TRUE for fasta, FALSE for RFP)}

\item{expression.file}{A string containing the location of a file containing
empirical expression rates, if needed.}

\item{append}{A boolean value that states whether the genes should be appended
 to the end of the genome}
}
\value{
This function returns the Genome Rcpp object created.
}
\description{
\code{initializeGenomeObject} initializes the Rcpp Genome object
}

