% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genomeObject.R
\name{length.Rcpp_Genome}
\alias{length.Rcpp_Genome}
\title{Length of Genome}
\usage{
\method{length}{Rcpp_Genome}(genome)
}
\arguments{
\item{genome}{A genome object initialized with \code{\link{initializeGenomeObject}}.}
}
\value{
returns the number of genes in a genome
}
\description{
\code{length} gives the length of a genome
}

