% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{loadParameterObject}
\alias{loadParameterObject}
\title{Load Parameter Object}
\usage{
loadParameterObject(file)
}
\arguments{
\item{file}{A filename that where the data will be stored.}

\item{parameter}{A parameter object that corrosponds to
one of the model types, such as "ROC", or "FONSE".}

\item{model}{Type of the model. Should corrospond to the parameter type.}
}
\value{
This function has no return value.
}
\description{
\code{loadParameterObject} will call the appropriate followup
call to loadXXXParameterObject based off of the parameter type 
given.
}
\details{
For example, if a ROCParameter is passed, the the loadParameterObject
for the ROCParameter will be called. This allows us to not have an if-else
block in the code - making use of the how R handles these situations.
}

