% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{writeParameterObject.Rcpp_RFPParameter}
\alias{writeParameterObject.Rcpp_RFPParameter}
\title{Write RFP Parameter Object to a File}
\usage{
\method{writeParameterObject}{Rcpp_RFPParameter}(parameter, file)
}
\arguments{
\item{parameter}{A parameter object that corrosponds to
the RFP model.}

\item{file}{A filename that where the data will be stored.
The file should end with the extension "Rdat".}
}
\value{
This function has no return value.
}
\description{
\code{writeRFPParameterObject} saves RFP parameter information 
to a file which can be used to re-initialize the object.
}
\details{
The file is written via R's save function. This function should
never be called; instead, \code{writeParameterObject} should be called
and allowed to make the necessary function call.
}

