% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{writeParameterObject}
\alias{writeParameterObject}
\title{writeParameterObject}
\usage{
writeParameterObject(parameter, file)
}
\arguments{
\item{parameter}{A parameter object that corrosponds to
one of the model types, such as "ROC", or "FONSE".}

\item{file}{A filename that where the data will be stored.
The file should end with the extension "Rdat".}
}
\description{
This function will call the appropriate followup
call to writeParameterObject based off of the parameter type 
given. For example, if a ROCParameter is passed, the the writeParameterObject
for the ROCParameter will be called. This allows us to not have an if-else
block in the code.
}

