% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{writeParameterToCSV.Rcpp_ROCParameter}
\alias{writeParameterToCSV.Rcpp_ROCParameter}
\title{Write ROC Parameter To CSV File}
\usage{
\method{writeParameterToCSV}{Rcpp_ROCParameter}(parameter, filename = NULL,
  CSP = NULL, mixture = 1, samples = 10)
}
\arguments{
\item{parameter}{A parameter object whose type should be "ROC".}

\item{filename}{A filename where the data will be written to.
This file should end with a "csv" extension. The default value 
is NULL.}

\item{CSP}{Tells what codon specific parameter should be written to the file.
"Mutation" and "Selection" are the valid arguments. The default value is NULL.}

\item{mixture}{Tells which mixture the data should be retrieved from to write.
The default value is 1.}

\item{samples}{The number of samples that should be used when calculating
the posteriors. The default value is 10.}
}
\value{
This function has no return value.
}
\description{
\code{writeROCParameterToCSV} will obtain the codon specific
parameter data for a given parameter and mixture and write this data
to a csv file. This function should never be called. \code{writeParameterToCSV}
should be called instead.
}
\details{
\code{writeROCParameterToCSV} For the given CSP, the posterior mean will
be calculated based off the number of samples for each amino acid given. If no
filename is given, the results will be printed to the screen.
}

