% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{writeParameterToCSV}
\alias{writeParameterToCSV}
\title{Write Parameter To CSV File}
\usage{
writeParameterToCSV(parameter, filename, CSP, mixture, samples)
}
\arguments{
\item{parameter}{A parameter object that corrosponds to
one of the model types. Valid values are "ROC", "RFP", and
"FONSE".}

\item{filename}{A filename where the data will be written to.
This file should end with a "csv" extension.}

\item{CSP}{Tells what codon specific parameter should be written to the file.
This will vary between models.}

\item{mixture}{Tells which mixture the data should be retrieved from to write.}

\item{samples}{The number of samples that should be used when calculating
the posteriors.}
}
\value{
This function has no return value.
}
\description{
\code{writeParameterToCSV} will obtain the codon specific
parameter data for a given parameter and mixture and write this data
to a csv file.
}
\details{
\code{writeParameterToCSV} will make the necessary calls
to writeXXXParameterToCSV based off of the parameter given.
}

