% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterObject.R
\name{writeParameterObject}
\alias{writeParameterObject}
\title{Write Parameter Object to a File}
\usage{
writeParameterObject(parameter, file)
}
\arguments{
\item{parameter}{A parameter object that corrosponds to
one of the model types, such as "ROC", or "FONSE".}

\item{file}{A filename that where the data will be stored.
The file should end with the extension "Rdat".}
}
\value{
This function has no return value.
}
\description{
\code{writeParameterObject} will call the appropriate followup
call to writeXXXParameterObject based off of the parameter type 
given.
}
\details{
For example, if a ROCParameter is passed, the the writeParameterObject
for the ROCParameter will be called. This allows us to not have an if-else
block in the code - making use of the how R handles these situations.
}

